feat(bubu-log): only enforce login consent in production
